pyproject.toml
[tool.poetry]
name = "python-utilities"
version = "0.0.1"
description = "Centralized python package to handle basic python operations"
authors = ["Tahn Thawainin <pzt002@auburn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
pandas = "^2.2.1"
pyyaml = "^6.0.1"
myaml = "^1.0.1"
python-box = "^7.1.1"
matplotlib = "^3.8.3"
pytest = "^8.1.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
